Update 22 N num.c

+++ 22 N num.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 
 int main(void) {
-           long int a[100],n,i,max=0;
-           scanf("%ld",&n);
+           int a[100],n,i,max=0;
+           scanf("%d",&n);
            for(i=0;i<n;i++)
-           scanf("%ld",&a[i]);
+           scanf("%d",&a[i]);
            max=a[0];
            for(i=0;i<n;i++)
            {
            	if(a[i]>max)
            	max=a[i];
            }
-           printf("%ld",max);
+           printf("%d",max);
 	return 0;
 }
